+++ Atomic/diff.py
@@ -4,6 +4,38 @@ import rpm
 import Atomic.util as util
 from filecmp import dircmp
 import Atomic.mount as mount
+from Atomic import Atomic
+
+class Diff(Atomic):
+    def diff(self):
+        '''
+        Allows you to 'diff' the RPMs between two different docker images|containers.
+        :return: None
+        '''
+        helpers = DiffHelpers(self.args)
+        images = self.args.compares
+        # Check to make sure each input is valid
+        for image in images:
+            self.get_input_id(image)
+
+        image_list = helpers.create_image_list(images)
+
+        # Set up RPM classes and make sure each docker object
+        # is RPM-based
+        if self.args.rpms:
+            rpm_image_list = helpers.build_rpm_list(image_list)
+
+        if not self.args.no_files:
+            helpers.output_files(images, image_list)
+
+        if self.args.rpms:
+            helpers.output_rpms(rpm_image_list)
+
+        # Clean up
+        helpers._cleanup(image_list)
+
+        if self.args.json:
+            util.output_json(helpers.json_out)
 
 class DiffHelpers(object):
     """
